Update JSON schemas to standardize on raw.githubusercontent.com
This was chosen over the static hosted option as this url would be more
generally reachable and usable direct from the source without needing to
go through a source compilation step. There may be times where the github
pages go down, so we want to make sure we leverage something closer to
the source to minimize dependency chain issues.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>