Implement version publishing utilities
We need to update the paths inside versioned documentation copies to make
sure all assets are properly resolvable. This script does this by making
all files that match a certain pattern get updated to the right format
and version. This must be done prior to doing a formal version release.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>