Change system_id to use shorthand ITSM identification
While URIs may be more canonical in nature, in practice they introduce
unnecessary overhead when they need to be URIEncoded in a query parameter.
As we do not have any anticipated systems that are outside of ITSM, we
will run with the preliminary assumption that all systems will be tracked
by ITSM. To simplify system multiplicities, we also remove the one to many
optionality for record identification.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>